remove Provider.WorkerSize and add workerSize argument to Zone()
This is not ideal, as only the AWS implementation needs to know
about WorkerSize. This means that WorkerSize should be passed only
when we're constructing an AWSProvider; however there's a cyclic
dependency between Provider and Config - we need some Config to set
Provider, and we need some of Provider's functionality to set Config.

This change means we no longer need the AWS-only WorkerSize method,
but means that all callers of Zone() now need to provide a
meaningless argument.

The right answer is to separate out the interdependencies between
Provider and Config, so that WorkerSize can be provided upon
construction of AWSProvider as described above; then we can remove
the additional argument to Zone().