Consistent "Name" tag (with uppercase N)
AWS resource tags are case sensitive.  The "Name" tag is pretty standard, and widely used, but some of the control-tower resources have a "name" tag, which is confusing, especially as many screens in the AWS console automatically show the "Name" tag but not the "name" tag.  This commit consistently tags resources with the "Name" tag.