Turn on EBS disk encryption at CPI level
We're already hardcoding disk encryption on all the vm types in the
cloud config. This change enforces it on the director as well as on root
and ephemeral disks.

Changes made here are consistent with the encrypted-disk ops file in
bosh-deployment: https://github.com/cloudfoundry/bosh-deployment/blob/master/aws/encrypted-disk.yml