Reserve the 7th IP in the public cidr on AWS
The NAT gateway gets placed on this IP which makes it unavailable for
BOSH-managed VMs. This doesn't matter for vanilla CT deployments since
the web is coded to be placed on the 8th IP but if you then use the
director to deploy other things you get vm placement errors.