Use specific user for AWS self update instead of deployer's creds
Our initial implementation of the self update pipeline just pulls the
AWS creds from the session of the most recent manual deploy and writes
them directly into the pipeline. This is very insecure as it means
anyone with fly get-pipeline access to the main team can see the creds.

This commit makes the following changes:
- Use terraform to create a special user who's keys will be used in the
  self update pipeline
- Writes the keys of the new user into credhub after Concourse has been
  deployed so that they don't appear in fly get-pipeline output
- Slims down the iam permissions of the user (I tried to figure out
  slimmed down s3 and ec2 as well but it got complicated so I'll leave
  that to a future update)
- Limits the new user's permissions to only work when requests originate
  from the Concourse's NAT IP (i.e. from a Concourse worker)